HBX-1679: Move class 'org.hibernate.tool.hbm2x.HibernateMappingGlobalSettings' to 'org.hibernate.tool.internal.export.hbm.HibernateMappingGlobalSettings'
