+++ orm/src/main/java/org/hibernate/tool/internal/export/hbm/HibernateMappingExporter.java
@@ -11,7 +11,6 @@ import org.hibernate.boot.Metadata;
 import org.hibernate.internal.util.StringHelper;
 import org.hibernate.mapping.PersistentClass;
 import org.hibernate.tool.hbm2x.GenericExporter;
-import org.hibernate.tool.hbm2x.HibernateMappingGlobalSettings;
 import org.hibernate.tool.hbm2x.TemplateProducer;
 import org.hibernate.tool.internal.export.pojo.POJOClass;
